CMark

Documentation

commonmarkToHtml :: [CMarkOption] -> Text -> Text

Convert CommonMark formatted text to Html, using cmark's built-in renderer.

commonmarkToXml :: [CMarkOption] -> Text -> Text

Convert CommonMark formatted text to CommonMark XML, using cmark's built-in renderer.

commonmarkToMan :: [CMarkOption] -> Maybe Int -> Text -> Text

Convert CommonMark formatted text to groff man, using cmark's built-in renderer.

commonmarkToLaTeX :: [CMarkOption] -> Maybe Int -> Text -> Text

Convert CommonMark formatted text to latex, using cmark's built-in renderer.

commonmarkToNode :: [CMarkOption] -> Text -> Node

Convert CommonMark formatted text to a structured Node tree, which can be transformed or rendered using Haskell code.

nodeToHtml :: [CMarkOption] -> Node -> Text

nodeToXml :: [CMarkOption] -> Node -> Text

nodeToMan :: [CMarkOption] -> Maybe Int -> Node -> Text

nodeToLaTeX :: [CMarkOption] -> Maybe Int -> Node -> Text

nodeToCommonmark :: [CMarkOption] -> Maybe Int -> Node -> Text

optSourcePos :: CMarkOption

Include a data-sourcepos attribute on block elements.

optNormalize :: CMarkOption

Normalize the document by consolidating adjacent text nodes.

optHardBreaks :: CMarkOption

Render softbreak elements as hard line breaks.

optSmart :: CMarkOption

Convert straight quotes to curly, --- to em-dash, -- to en-dash.

optSafe :: CMarkOption

Suppress rendering of raw HTML and potentially dangerous URLs in links and images.

optUnsafe :: CMarkOption

Allow rendering of raw HTML and potentially dangerous URLs in links and images.

data Node

Constructors

Node (Maybe PosInfo) NodeType [Node]

data NodeType

Constructors

DOCUMENT

THEMATIC_BREAK

PARAGRAPH

BLOCK_QUOTE

HTML_BLOCK Text

CUSTOM_BLOCK OnEnter OnExit

CODE_BLOCK Info Text

HEADING Level

LIST ListAttributes

ITEM

TEXT Text

SOFTBREAK

LINEBREAK

HTML_INLINE Text

CUSTOM_INLINE OnEnter OnExit

CODE Text

EMPH

STRONG

LINK Url Title

IMAGE Url Title

data PosInfo

Constructors

PosInfo

Fields

startLine :: Int

startColumn :: Int

endLine :: Int

endColumn :: Int

data DelimType

Constructors

PERIOD_DELIM

PAREN_DELIM

data ListType

Constructors

BULLET_LIST

ORDERED_LIST

data ListAttributes

Constructors

ListAttributes

Fields

listType :: ListType

listTight :: Bool

listStart :: Int

listDelim :: DelimType

type Url = Text

type Title = Text

type Level = Int

type Info = Text

data CMarkOption